solana-service/requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.6
aiohttp>=3.9.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
ansi2html==1.9.2
anyio==4.8.0
apischema==0.17.5
APScheduler>=3.10.4
asyncio>=3.4.3
attrs==25.1.0
audioop-lts==0.2.1
Authlib==0.15.6
base58==2.1.1
based58==0.1.1
beautifulsoup4==4.13.3
bitarray==3.1.0
blinker==1.9.0
cachetools==4.2.2
ccxt==4.2.14
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==1.0.2
click==8.1.8
construct==2.10.68
construct-typing==0.5.6
cryptography==44.0.1
Cython==3.0.11
cytoolz==1.0.1
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.5.9
dateparser>=1.2.1
decorator==5.2.1
discord.py>=2.3.2
eth-account==0.11.3
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==1.0.1
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
filelock==3.17.0
Flask==3.0.0
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
hexbytes==0.3.1
httpcore>=0.15.0,<0.17.0
httpx>=0.23.0,<0.24.0
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonalias==0.1.1
jsonify==0.5
jsonrpcclient==4.0.3
jsonrpcserver==5.0.9
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lru-dict==1.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-enum==1.5.1
mdurl==0.1.2
millify==0.1.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.27.1
nest-asyncio==1.6.0
networkx==3.4.2
newsapi-python==0.2.7
nltk==3.9.1
nodejs==0.1.1
numpy>=1.24.0
oauthlib==3.2.2
optional-django==0.1.0
OSlash==0.6.3
packaging==24.2
pandas>=2.0.0
parsimonious==0.10.0
pillow==11.1.0
plotly==5.18.0
praw==7.7.1
prawcore==2.4.0
prettytable==3.14.0
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
pyaes==1.6.1
pyarrow==19.0.1
pyasn1==0.6.1
pycares==4.5.0
pycoingecko==3.2.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyecharts==2.0.8
Pygments==2.19.1
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-telegram-bot>=20.7
python-twitter==3.5
python-twitter-v2==0.8.1
pytz>=2024.2
pyyaml==6.0.1
pyunormalize==16.0.0
ratelimit==2.2.1
redis==5.0.1
referencing==0.36.2
regex==2024.11.6
requests>=2.31.0
requests-oauthlib>=1.2.0,<2.0.0
retrying==1.3.4
rfc3986==1.5.0
rich==13.9.4
rlp==4.1.0
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.2
scikit-learn==1.4.0
scipy==1.15.2
sentry-sdk==1.39.1
setuptools==75.8.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
streamlit==1.42.2
streamlit-aggrid==1.1.0
streamlit-autorefresh==1.0.1
streamlit-echarts==0.4.0
streamlit-option-menu==0.4.0
streamlit-toggle-switch==1.0.2
sympy==1.13.1
ta==0.11.0
ta-lib==0.6.3
Telethon==1.39.0
tenacity==8.2.3
textblob==0.19.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
toolz>=0.11.2,<0.12.0
torch==2.6.0
tornado==6.1
tqdm==4.67.1
transformers==4.49.0
tweepy>=4.14.0
types-cachetools==4.2.10
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.3
ujson==5.10.0
update-checker==0.18.0
urllib3==2.3.0
vaderSentiment==3.3.2
wcwidth==0.2.13
web3==6.15.1
websocket-client==1.8.0
websockets>=10.0.0,<12.0
Werkzeug==3.0.1
wheel==0.45.1
yarl==1.18.3
zipp==3.21.0

# Solana Dependencies
solana==0.32.0
anchorpy==0.19.0
solders==0.20.0

# Telegram Bot
python-telegram-bot>=20.7
